Add new car model :car:
New car model and prefab that looks nicer and contains an interior. This
also includes a steering wheel which moves as the user does. This will
have to be configured to match the wheel if possible.

Car model was taken from:
https://free3d.com/3d-model/bmw-m5-69595.html
But has been modified to better fit the purpose (i.e. fill interior with
more detail and put steering wheel on opposite side).