Fixed issue with VR framerates and aligneds position.
The framerate issue was caused by potentially a few issues: having
multiple game windows open at the same time, having multiple displays
plugged in, or having too many errors at the same time. All of these
were fixed at this commit, but not all were tried individually.

The deadzone of the controller was removed, and the wheel turn was
aligned with the horizontal movement.